refactor: migrate markdown parser tests from text editor and viewer component to model scripts
